# server/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.JobRequest:
    properties:
      applied_at:
        type: string
      company:
        type: string
      interviewed_at:
        type: string
      job_url:
        type: string
      location_type:
        type: string
      notes:
        type: string
      position:
        type: string
      salary:
        type: string
      source:
        type: string
      status:
        type: string
    type: object
  handler.JobResponse:
    properties:
      applied_at:
        type: string
      company:
        type: string
      interviewed_at:
        type: string
      job_url:
        type: string
      location_type:
        type: string
      notes:
        type: string
      position:
        type: string
      salary:
        type: string
      source:
        type: string
      status:
        type: string
    type: object
  handler.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  handler.LoginResponse:
    properties:
      token:
        type: string
      username:
        type: string
    type: object
  handler.RegisterRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: contact@jesstin.com
    name: Jesstin Swadley
    url: https://jesstin.com/support
  description: This is a job tracking server.
  title: Job Tracker API
  version: "1.0"
paths:
  /jobs:
    get:
      description: Retrieve a list of all job applications for the logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.JobResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all jobs
      tags:
      - jobs
    post:
      consumes:
      - application/json
      parameters:
      - description: Job Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.JobRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.JobResponse'
      security:
      - BearerAuth: []
      summary: Create a new job
      tags:
      - jobs
  /jobs/{id}:
    delete:
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete a job
      tags:
      - jobs
    put:
      consumes:
      - application/json
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Job Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.JobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.JobResponse'
      security:
      - BearerAuth: []
      summary: Update a job
      tags:
      - jobs
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return a JWT token
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User Login
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Create a new user account for the job tracker
      parameters:
      - description: Registration Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.LoginResponse'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: User already exists
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer ' followed by your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"